Refactor RedisMessageType enum
